Extract NftsCarousel to its own component
